Refactor structure of settings npo application page
